pyproject.toml
[tool.poetry]
name = "flowershop"
version = "0.1.0"
description = ""
authors = ["Stranix <Phantom2525@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
environs = "^9.5.0"
gunicorn = "^21.2.0"
dj-database-url = "^2.0.0"
pillow = "^10.0.0"
python-telegram-bot = "13.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
